create vehicle lookup component feature
