+++ ADP.WebComponents/adp-web-components/src/features/vehicle-lookup-component/docs.md
@@ -1,0 +1,54 @@
+## static content ( can be extended )
+
+```typescript
+
+  // ====== Start Vehicle Lookup Component Shared Logic
+
+ @Prop() isDev: boolean;
+  @Prop() baseUrl: string;
+  @Prop() headers: object = {};
+  @Prop() queryString: string = '';
+
+  @Prop() errorCallback?: (errorMessage: ErrorKeys) => void;
+  @Prop() loadingStateChange?: (isLoading: boolean) => void;
+  @Prop() loadedResponse?: (response: VehicleLookupDTO) => void;
+
+  @State() isError: boolean = false;
+  @State() errorMessage?: ErrorKeys;
+  @State() isLoading: boolean = false;
+  @State() vehicleLookup?: VehicleLookupDTO;
+
+  @Element() el: HTMLElement;
+
+  mockData;
+
+  abortController: AbortController;
+  networkTimeoutRef: ReturnType<typeof setTimeout>;
+
+  @Method()
+  async setMockData(newMockData: MockJson<VehicleLookupDTO>) {
+    this.mockData = newMockData;
+  }
+
+  @Method()
+  async setData(newData: VehicleLookupDTO | string, headers: any = {}) {
+    await setVehicleLookupData(this, newData, headers);
+  }
+
+  @Method()
+  async setErrorMessage(message: ErrorKeys) {
+    setVehicleLookupErrorState(this, message);
+  }
+
+  @Watch('isLoading')
+  onLoadingChange(newValue: boolean) {
+    if (this.loadingStateChange) this.loadingStateChange(newValue);
+  }
+
+  @Method()
+  async fetchData(requestedVin: string, headers: any = {}) {
+    await this.setData(requestedVin, headers);
+  }
+
+ // ====== End Vehicle Lookup Component Shared Logic
+```
