+++ ADP.WebComponents/adp-web-components/src/features/vehicle-lookup-component/vehicle-lookup-api-integration.ts
@@ -1,7 +1,9 @@
 import { ErrorKeys } from '~features/multi-lingual';
 
 import { VehicleLookupDTO } from '~types/generated/vehicle-lookup/vehicle-lookup-dto';
-import VehicleLookupComponent, { vehicleRequestHeaders } from '~types/interfaces/vehicle-lookup-component';
+
+import { VehicleLookupComponent } from './interface';
+import { vehicleRequestHeaders } from './types';
 
 export const setVehicleLookupData = async (
   context: VehicleLookupComponent,
